Refactor input validation rule for settlement method
