+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/ValidationErrors/SettlementMethodMissingRequiredDomainValuesValidationError.cs
@@ -18,13 +18,10 @@ namespace Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErro
 {
     public class SettlementMethodMissingRequiredDomainValuesValidationError : ValidationError
     {
-        public SettlementMethodMissingRequiredDomainValuesValidationError(string gsrnNumber, string settlementMethod)
+        public SettlementMethodMissingRequiredDomainValuesValidationError(string settlementMethod)
         {
-            GsrnNumber = gsrnNumber;
             SettlementMethod = settlementMethod;
         }
-
-        public string GsrnNumber { get; }
 
         public string SettlementMethod { get; }
     }
