+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/CreateMeteringPoints/ConsumptionMeteringPointTests.cs
@@ -86,6 +86,20 @@ namespace Energinet.DataHub.MeteringPoints.IntegrationTests.CreateMeteringPoints
             AssertValidationError<CreateMeteringPointRejected>("E86");
         }
 
+        [Fact]
+        public async Task Should_reject_when_settlement_method_is_missing()
+        {
+            var request = CreateRequest()
+                with
+                {
+                    SettlementMethod = string.Empty,
+                };
+
+            await _mediator.Send(request).ConfigureAwait(false);
+
+            AssertValidationError<CreateMeteringPointRejected>("D02");
+        }
+
         private static CreateMeteringPoint CreateRequest()
         {
             return new CreateMeteringPoint(
